extract commit list (into TreeCommits) from TreePage
TreePage was getting big. There is no behavior or UI change, this is just a refactor.
