+++ client/web/src/repo/tree/TreePage.module.scss
@@ -17,19 +17,6 @@
     max-width: var(--media-xl);
 }
 
-.git-commit-node {
-    padding-left: 0;
-    padding-right: 0;
-
-    &-message-subject {
-        opacity: 0.9;
-    }
-
-    :global(.btn) {
-        opacity: 0.85;
-    }
-}
-
 .container {
     overflow-y: auto;
     height: 100%;
